% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contRibution.R
\docType{package}
\name{contRibution}
\alias{contRibution}
\title{contRibution: Visual documentation of project contributions}
\description{
Implements generation of a table to document author contributions to
research papers and projects as an alternative to paragraphs describing the
same. This replaces wordy descriptions with a visual and intuitive
standard.
}
