% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{contribution_table}
\alias{contribution_table}
\title{Generates an author contribution table}
\usage{
contribution_table(df)
}
\arguments{
\item{df}{data frame containing author contribution information
in two columns: Author, Activity. See example for details.}
}
\value{
ggplot of author contribution table
}
\description{
Generates an author contribution table
}
\examples{
# NK participated in Concept and Design
# RW wrote code
# JS participated in Concept and Design
info <- data.frame(Author = c("NK", "NK", "RW", "JS", "JS"),
                   Activity = c("Concept", "Design", "Code", "Concept",
                                "Design"))
# grey implies author participation in activity; white does not
contribution_table(info)
}
