% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo}
\alias{demo}
\title{A demo for plotting contribution table}
\format{A data.frame}
\source{
See data_raw directory
}
\description{
A demo for plotting contribution table
}
\examples{
data("demo")
}
