% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palette}
\alias{palette}
\title{palette}
\format{
A data.frame
}
\source{
See \url{https://github.com/williambelle/github-contribution-color-graph}
}
\description{
palette
}
\examples{
data("palette")
}
