% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_palette.R
\name{show_palette}
\alias{show_palette}
\title{Show supported palette}
\usage{
show_palette()
}
\value{
\code{NULL}
}
\description{
A modified version of \href{https://github.com/tyluRp/lisa/blob/master/R/utils.R}{plot.lisa_palette}.
}
\examples{
show_palette()
}
