% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generate contribution table}
\usage{
generate(data, color_map = c("white", "grey", "black"),
  show_legend = FALSE, title = NULL, xlab = NULL, ylab = NULL,
  caption = NULL, tag = NULL, font_size_x = 16, font_size_y = 16,
  text_angle_x = 30, coord_ratio = 1)
}
\arguments{
\item{data}{a \code{data.frame}. e.g. \code{data("demo")}.}

\item{color_map}{color map, either a length-3 vector for 3 contribution level:
None, Some and Many; or a \code{Scale} object like \code{scale_fill_brewer(palette ="Oranges")}.}

\item{show_legend}{if \code{TRUE}, show figure legend.}

\item{title}{The text for the title.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{caption}{The text for the caption which will be displayed in the
bottom-right of the plot by default.}

\item{tag}{The text for the tag label which will be displayed at the
top-left of the plot by default.}

\item{font_size_x}{font size for x.}

\item{font_size_y}{font size for y.}

\item{text_angle_x}{text angle for x.}

\item{coord_ratio}{coordinate ratio.}
}
\value{
a \code{ggplot2} object
}
\description{
Generate contribution table
}
\examples{
library(contribution)
library(ggplot2)
generate(demo)
generate(demo, text_angle_x = 20, color_map = scale_fill_brewer(palette ="Oranges"))
}
